+++ dl-api-csharp/Doublelives.Api/Models/Menu/ListFouRouterViewModel.cs
@@ -13,15 +13,15 @@ namespace Doublelives.Api.Models.Menu
 
         public bool Hidden { get; set; }
 
-        public long Id { get; set; }
+        public int Id { get; set; }
 
         public Meta Meta { get; set; }
 
         public string Name { get; set; }
 
-        public long Num { get; set; }
+        public int Num { get; set; }
 
-        public long ParentId { get; set; }
+        public int ParentId { get; set; }
 
         public string Path { get; set; }
     }
