+++ dl-api-csharp/Doublelives.Api/Models/Notice/NoticeViewModel.cs
@@ -9,18 +9,18 @@ namespace Doublelives.Api.Models.Notice
     {
         public string Content { get; set; }
 
-        public long CreateBy { get; set; }
+        public int CreateBy { get; set; }
 
         public DateTimeOffset CreateTime { get; set; }
 
-        public long Id { get; set; }
+        public int Id { get; set; }
 
-        public long ModifyBy { get; set; }
+        public int ModifyBy { get; set; }
 
         public DateTimeOffset ModifyTime { get; set; }
 
         public string Title { get; set; }
 
-        public long Type { get; set; }
+        public int Type { get; set; }
     }
 }
