+++ dl-api-csharp/Doublelives.Migrations/DbFirstProcess/MessageTemplate.cs
@@ -11,7 +11,7 @@ namespace Doublelives.Migrations.DbFirstProcess
 
         public string Content { get; set; }
 
-        public long IdMessageSender { get; set; }
+        public int IdMessageSender { get; set; }
 
         public string Title { get; set; }
