+++ dl-api-csharp/Doublelives.Migrations/DbFirstProcess/SysTaskLog.cs
@@ -9,7 +9,7 @@ namespace Doublelives.Migrations.DbFirstProcess
 
         public int? ExecSuccess { get; set; }
 
-        public long? IdTask { get; set; }
+        public int? IdTask { get; set; }
 
         public string JobException { get; set; }
